refactor(fragment): move netcdf_fragment
