+++ goesdr/data_fragment.py
@@ -12,8 +12,8 @@ DataFragment
     Represents a data record extracted from a netCDF dataset.
 """
 
-from inspect import currentframe
-from typing import Any, NoReturn
+from inspect import currentframe, getmodule
+from typing import Any, NoReturn, TypeVar, cast
 
 from netCDF4 import Dataset  # pylint: disable=no-name-in-module
 
@@ -155,3 +155,30 @@ class DataFragment:
                 f"Attribute '{name}: {atype}' "
                 f"does not match the given type '{vtype}'"
             )
+
+
+_T = TypeVar("_T")
+
+
+def netcdf_fragment(recordclass: type[_T]) -> type[_T]:
+    class _FragmentRecord(DataFragment):
+        pass
+
+    _FragmentRecord.__annotations__ = recordclass.__annotations__.copy()
+    _FragmentRecord.__module__ = recordclass.__module__
+    _FragmentRecord.__name__ = recordclass.__name__
+    _FragmentRecord.__qualname__ = recordclass.__qualname__
+
+    for name in recordclass.__annotations__:
+        if hasattr(recordclass, name):
+            value = getattr(recordclass, name)
+            setattr(_FragmentRecord, name, value)
+
+    __post_init__ = getattr(recordclass, "__post_init__", None)
+    if callable(__post_init__):
+        setattr(_FragmentRecord, "__post_init__", __post_init__)
+
+    original_module = getmodule(recordclass)
+    setattr(original_module, recordclass.__name__, _FragmentRecord)
+
+    return cast(type[_T], _FragmentRecord)
